Information and quote section for the landing page
The information section contains a title and flexed div. Within the div,
there are four image/illustration placeholders with text below them.
These placeholders and their respective text are housed in a vertical
flex box.

The quote section is a vertical flex box that houses a paragraph and h6
element. The paragraph is centered on both axes and the h6 element is
aligned to the left of the paragrah.