Call to action section and footer
The call to action section is a flex with three direct childeren, a
header, paragraph and container div for the button. The footer is simply
text on a background. Changing the font-family to roboto affected the
position of some items, particulary the navigation list. I increased
the width of the nav list container because the text within was wrapping
due to limitted space.